% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMTEssence.R
\name{CausalImpactBackTest}
\alias{CausalImpactBackTest}
\alias{causal_impact_test}
\title{my title}
\usage{
causal_impact_test(data, metric, pairs, pre_period, post_period,
  p_value_thresh, plots = F)
}
\arguments{
\item{data}{a data frame}

\item{metric}{conversion metric}

\item{pairs}{a list of market pairs}

\item{pre_period}{vector of two dates representing start and end of pre period}

\item{post_period}{vector of two dates representing start and end of post period}

\item{p_value_thresh}{the minumum p-value acceptable for valid test}

\item{plots}{Boolean for plot inclusion}
}
\value{
list of valid tests
}
\description{
my title
}
