% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMTEssence.R
\name{GetMarketMatches}
\alias{GetMarketMatches}
\alias{get_market_matches}
\title{my title}
\usage{
get_market_matches(data, metric, pre_period, post_period, corr_thresh,
  p_value_thresh, label = NULL, rank = 20, plots = F)
}
\arguments{
\item{data}{a data frame}

\item{metric}{conversion metric}

\item{pre_period}{vector of two dates representing start and end of pre period}

\item{post_period}{vector of two dates representing start and end of post period}

\item{corr_thresh}{the minimum correlation value for acceptable matches}

\item{p_value_thresh}{the minumum p-value acceptable for valid test}

\item{label}{a label for your MMT e.g. "YouTube TV Q1 2021"}

\item{rank}{the amount of matches for each DMA to look at, default is all}

\item{plots}{Boolean for plot inclusion}
}
\value{
list of valid matches
}
\description{
my title
}
